backend-functions/requirements.txt
azure-functions
azure-storage-blob
pandas>=2.2.0
numpy>=1.26.0
